Lazy-load editor panels and optimize project list
